Add support for proxy with authorization.
Changes:
- Import open-uri from Ruby's CVS HEAD
- Added proxy authorization discovery from proxy URL or from HTTP_PROXY_USER
  and HTTP_PROXY_PATH environment variables.  Includes tests.
- Fix Rubygems tests on Windows.


git-svn-id: svn+ssh://rubyforge.org/var/svn/rubygems/trunk@1008 3d4018f9-ac1a-0410-99e9-8a154d859a19

+++ rubygems/test/test_gempaths.rb
@@ -109,6 +109,8 @@ class TestGemPaths < RubyGemTestCase
   end
 
   def test_ensure_gem_directories_write_protected
+    return if win_platform? #This test works only for FS that support write protection
+    
     gemdir = File.join(@tempdir, "egd")
     FileUtils.rm_r gemdir rescue nil
     FileUtils.mkdir_p gemdir
@@ -122,6 +124,8 @@ class TestGemPaths < RubyGemTestCase
   end
 
   def test_ensure_gem_directories_with_parents_write_protected
+    return if win_platform? #This test works only for FS that support write protection
+
     parent = File.join(@tempdir, "egd")
     gemdir = "#{parent}/a/b/c"
     
@@ -130,7 +134,7 @@ class TestGemPaths < RubyGemTestCase
     FileUtils.chmod 0400, parent
     Gem.use_paths(gemdir)
     Gem.send(:ensure_gem_subdirectories, gemdir)
-    assert ! File.exist?("#{gemdir}/cache")
+    assert !File.exist?("#{gemdir}/cache")
   ensure
     FileUtils.chmod(0600, parent) rescue nil
     FileUtils.rm_r parent rescue nil
